# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for the Noterian note-taking application.\nAuthentication is performed via a JWT-cookie set by /signup and /signin.\nState-changing endpoints additionally require an X-CSRF-Token header (obtain it from /csrf-token).",
        "title": "WHITECROWSOFT API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/notes/{noteId}/attachments": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "Загружает файл в MinIO и создаёт блок-аттач в заметке. Поддерживает image, gif, audio, video.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "Загрузить аттач в заметку",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID заметки",
                        "name": "noteId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Позиция вставки блока",
                        "name": "position",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "Файл вложения",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-park-mail-ru_2026_1_WHITECROWSOFT_internal_attachments_dto.Attachment"
                        }
                    },
                    "400": {
                        "description": "Недопустимый mime-type или позиция",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Заметка или блок не найдены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "У блока уже есть аттач",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "413": {
                        "description": "Файл слишком большой",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/notes/{noteId}/blocks/{blockId}/attachments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "Получить аттач блока",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID заметки",
                        "name": "noteId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID блока",
                        "name": "blockId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-park-mail-ru_2026_1_WHITECROWSOFT_internal_attachments_dto.Attachment"
                        }
                    },
                    "400": {
                        "description": "Некорректный noteId/blockId",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Заметка, блок или аттач не найдены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "Удалить аттач",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID заметки",
                        "name": "noteId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID блока",
                        "name": "blockId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Аттач удалён"
                    },
                    "400": {
                        "description": "Некорректный noteId/blockId",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Заметка, блок или аттач не найдены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_go-park-mail-ru_2026_1_WHITECROWSOFT_internal_attachments_dto.Attachment": {
            "type": "object",
            "properties": {
                "attach_url": {
                    "type": "string"
                },
                "block_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "minio_key": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url_expires_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT-токен сессии, выставляется сервером после /signup или /signin.",
            "type": "apiKey",
            "name": "token",
            "in": "cookie"
        },
        "CsrfToken": {
            "description": "CSRF-токен, полученный из /csrf-token. Требуется для POST/PUT/DELETE.",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    },
...
}